Broken, Save
Keen broke something the the last updates, and the use of "Config" as a
directory name causes the game not to recognise files in that directory.
"Configuration" is not recognised either, so I've renamed it to
"EconConfig".

I have added code that catches when the game saves, so it is saved too.
I have removed the save on quit, as the SaveData() is called by the game
when exiting anyhow so it will be saved (It's stopping double saving),
except in a dedicated server that has autosave set to 0. The game
doesn't save when exiting under that condition, so I don't think we
should either, otherwise we risk getting unsynced.