Logging changes, Character detection
Generally, just enhancing the Logging. I've changed the logging to write
the Global storage path, instead of the Local, as logging in general
should be more visible. I've added a Debug to it, so if we are running
it locally, it will write a single log file, to enhance debugging. This
can be expanded later to write messages to screen also.
There is a IsExperimentalCreator, which SaltPepp should add their Steam
Id to if they want a better debugging experience.
A new method to find the player's character when in a cockpit to Remote
Control. Doesn't work when player is controlling a turret. It's a
kludge, and I've prompted Keen to do something about it.
And I've let my Visual Studio reorder all the using namespaces. My OCD
kicking in.