Restructure xml data
The number of files needed to define the xml structure is starting to
grow, so I've moved them into a separate directory and renamed the Xml
Element names to try and simplify them. This will break any current
saved data unfortunately.

A default Market for the NPC has been established in the centre of the
World, which contains a definition of the shape and size.