complete /deny /cancel
Code is untested:
Finished code for /Deny and /Cancel.
Fixed issue with accept leave order to be collected a second time if it
was put into inventory automatically.
Optimized the Lcd code to improve performance.
Moved a couple of the constants around, because these two are important
and shouldn't be buried.