#77
Most of the /worth functionality.
The only parts missing are gas calculations.
All inventory, and construction inventory, and partially built blocks
are accounted for. Strangely the method I used in SEToolbox is not
accurate. I looked through the Git code, and couldn't figure out how to
calculate partial blocks. Thankfully there is an API which tells me
exactly what is missing, and I deduct back.
I've done something unusual with this one, and that is run the
calculation in the background on the server. The concern is, running
/worth on a big ship, may impact the Sim Speed. By pushing it to another
thread, the Main Thread in which calculations occur, should be free to
do Physics calculations.
If this background processing works well, I'll expand it to the LCD
processing, as we could have more live information then. More testing
required to confirm.