#77
Added cockpit oxygen.
All gasses are classes as Inventory for the sake of calculations.

This command has some potential of crashing the server because of how
the background processing works. I'm not sure why, considering the Try
Catch methodology should keep it stable, but exceptions I tested with in
the code can cause the application to just crash without warning.