#40 server side Localization support
Added new Config setting to allow the Server to specify the language.
This affects:
The name of components on LCD.
The format of numbers and dates on LCD.
The name of components when messages are sent back to Clients.

Currently not affected:
The formats of values and dates getting sent back to Clients.

This does not affect:
The identification of components on the Client.