Fixed #100
Number of things addressed:
1. I've improved the logging, so it caches log entries instead of
writing them to disc immediately. Intensive I/O can cause slowdowns. It
will only write out to the log every 20th entry. The only issue with
this is, if the game crashes from an exception, we may not have all the
log entries in the file. and, you won't have all entries if you are
looking at the log as the game is running.
2. The code I wrote for interrogating through connected grids was broken
when pistons were involved.
3. I was using my own code for interrogating through connected grids for
buy and sell. It's not as fast as it could be, but there is an API that
can be used instead which is faster and sufficient for our needs.
4. I added a lot of logging to the buy and sell commands to track down
any further performance issues.