#63 readable price list
I was contemplating this a couple of weeks ago, and then forgot all
about it.
I'm re-boxing the code for managing the text into a new class, so it can
be used with Dialogs to format the text.
This makes the price list much better to read, and allows more
formatting.
I've pushed the executing of the pricelist to a background thread.
I'm not sure how this current SE netcode is supposed to run. Is running
complex calculations client side going to impact the server sim speed?
If so, then we're going to need to push more stuff to another thread.