#93 first additions
First set of working commands for adding, deleting and listing npc
market zones.
Creating and deleting of zones uses the Display Name. I'm going with
assumption that display names are Unique. There are checks for existing
market zones of the specified name.
Zone names could have "   " added to the end of a name. It's acceptable.
Yet to be added, the ability to rename a zone, and move an existing zone
(so the inventory is not lost).