#37 Added ClientConfig for client side data
ClientConfig will now handle client side data. Upon registration, the
Server will send the necessary packet of data for the client to know
what it needs.
Also, the player's BankBalance will routinely be updated by any
transactions performed.
This is yet to be linked to to automatically updated the Hud, which will
need an overhaul to modularlize it.
the previous Config property was renamed to ServerConfig to reduce
confusion.