Player market updates and restrictions
Added Configurable maximum limit to number of player markets, default to
1.
Added ability to open and close player markets.
Added ability to set buy and sell prices, and blacklist items.
Fixed so that upon setting item to blacklist in server config, all
markets are updated to prevent an item been sold on the server.