Changes reactive price logic so that buy never exceeds sell
When stock is too high, the buy from player price drops. But not the
sell to player price.
When stock is too low, the sell to price player goes up, but the buy
from player price doesnt.
Ideally the buy from player price should be allowed to increase UP to
the sell price but never exceed it
Conversely the sell to player price should be allowed to drop DOWN to
the buy from player price but never exceed it.   however   this logic
would require a way to know the price book buy/sell prices when
adjusting price.. this commit is a simple but effective alternate
solution for the short term