Bunch of paranoia fixes to avoid entity null reference crashes
apex / core systems update added extra checks to player/npc entity ID's to avoid clashing with existing players or space master IDs,  on lagged client or server this allows the possibility of stale MyAPIGateway.Session.Player entities slipping by for a few ticks each time the player respawns at a medbay; throwing a null reference crash to desktop when you try to update client objects against player entity.   These paranoia fixes halt or skip events that touch entities for a few seconds any time a player is detected to be respawning or dead as a work around to give the client time to catch up with server in these scanerios.